requirements.txt
ansible>=9,<11
requests>=2.31.0
PyYAML>=6.0
pexpect>=4.9.0
